backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.0
django-cors-headers==4.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
